+++ netpyne/cell/pointCell.py
@@ -45,9 +45,6 @@ class PointCell (Cell):
 
         # enable defining pointCell properties directly in cellParams
         exclude = ['cellType']
-
-        #import IPython; IPython.embed()
-
         if 'cellType' in tags and tags['cellType'] in sim.net.params.cellParams:
             tags.update({k: v for k,v in sim.net.params.cellParams[tags['cellType']].items() if k not in exclude})