+++ netpyne/network/pop.py
@@ -11,6 +11,7 @@ from __future__ import absolute_import
 
 from builtins import map
 from builtins import range
+from curses import has_key
 try:
     basestring
 except NameError:
@@ -144,8 +145,7 @@ class Pop (object):
         for i in self._distributeCells(numCells)[sim.rank]:
             gid = sim.net.lastGid+i
             self.cellGids.append(gid)  # add gid list of cells belonging to this population - not needed?
-            cellTags = {k: v for (k, v) in self.tags.items() if k in sim.net.params.popTagsCopiedToCells}  # copy all pop tags to cell tags, except those that are pop-specific
-            cellTags['pop'] = self.tags['pop']
+            cellTags = self._createCellTags()
             cellTags['xnorm'] = randLocs[i,0] # set x location (um)
             cellTags['ynorm'] = randLocs[i,1] # set y location (um)
             cellTags['znorm'] = randLocs[i,2] # set z location (um)
@@ -283,8 +283,7 @@ class Pop (object):
         for i in self._distributeCells(self.tags['numCells'])[sim.rank]:
             gid = sim.net.lastGid+i
             self.cellGids.append(gid)  # add gid list of cells belonging to this population - not needed?
-            cellTags = {k: v for (k, v) in self.tags.items() if k in sim.net.params.popTagsCopiedToCells}  # copy all pop tags to cell tags, except those that are pop-specific
-            cellTags['pop'] = self.tags['pop']
+            cellTags = self._createCellTags()
             cellTags['xnorm'] = randLocs[i,0]  # calculate x location (um)
             cellTags['ynorm'] = randLocs[i,1]  # calculate y location (um)
             cellTags['znorm'] = randLocs[i,2]  # calculate z location (um)
@@ -312,8 +311,7 @@ class Pop (object):
             #    self.cellModelClass = getattr(f, self.tags['cellsList'][i]['cellModel'])  # select cell class to instantiate cells based on the cellModel tags
             gid = sim.net.lastGid+i
             self.cellGids.append(gid)  # add gid list of cells belonging to this population - not needed?
-            cellTags = {k: v for (k, v) in self.tags.items() if k in sim.net.params.popTagsCopiedToCells}  # copy all pop tags to cell tags, except those that are pop-specific
-            cellTags['pop'] = self.tags['pop']
+            cellTags = self._createCellTags()
             cellTags.update(self.tags['cellsList'][i])  # add tags specific to this cells
             for coord in ['x','y','z']:
                 if coord in cellTags:  # if absolute coord exists
@@ -366,8 +364,7 @@ class Pop (object):
         for i in self._distributeCells(numCells)[sim.rank]:
             gid = sim.net.lastGid+i
             self.cellGids.append(gid)  # add gid list of cells belonging to this population - not needed?
-            cellTags = {k: v for (k, v) in self.tags.items() if k in sim.net.params.popTagsCopiedToCells}  # copy all pop tags to cell tags, except those that are pop-specific
-            cellTags['pop'] = self.tags['pop']
+            cellTags = self._createCellTags()
             cellTags['xnorm'] = gridLocs[i][0] / sim.net.params.sizeX # set x location (um)
             cellTags['ynorm'] = gridLocs[i][1] / sim.net.params.sizeY # set y location (um)
             cellTags['znorm'] = gridLocs[i][2] / sim.net.params.sizeZ # set z location (um)
@@ -379,6 +376,12 @@ class Pop (object):
         sim.net.lastGid = sim.net.lastGid + numCells
         return cells
 
+    def _createCellTags(self):
+        from .. import sim
+        # copy all pop tags to cell tags, except those that are pop-specific
+        cellTags = {k: v for (k, v) in self.tags.items() if k in sim.net.params.popTagsCopiedToCells}
+        cellTags['pop'] = self.tags['pop']
+        return cellTags
 
     def _setCellClass (self):
         """
@@ -414,6 +417,16 @@ class Pop (object):
                     self.tags['params'] = params
                     for k in self.tags['params']: self.tags.pop(k)
                     sim.net.params.popTagsCopiedToCells.append('params')
+
+                    # populate tags with params defined in `cellParams` globally for given cell type
+                    if 'cellType' in self.tags:
+                        type = self.tags['cellType']
+
+                        globalCellParams = { k: v for k,v in sim.net.params.cellParams[type].items() if k not in excludeTags }
+                        for k, v in globalCellParams.items():
+                            # but don't override more specific value from `popParams` (if exists) with more general from `cellParams`
+                            if k not in params:
+                                params[k] = v
             except:
                 if getattr(self.tags, 'cellModel', None) in ['NetStim', 'DynamicNetStim', 'VecStim', 'IntFire1', 'IntFire2', 'IntFire4']:
                     print('Warning: could not find %s point process mechanism required for population %s' % (cellModel, self.tags['pop']))