+++ CHANGES.md
@@ -13,6 +13,8 @@
 
 - Fixed bug with loading CompartCell with custom mechanisms from saved network
 
+- Fixed reading default properties for PointCell from cellParams
+
 # Version 1.0.2.1
 
 **New features**